Add header section for homepage
This replaces the current "MAINTENANCE MODE" page with a page consisting
of only the header section.

The newly-implemented section is a background video behind a darkened
overlay. On top of it, there are the SFU Summer Festival logo and the
"Enter" button.

This part is done desktop-first. It works best on a 1920x1080px screen.